requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
openai
requests
beautifulsoup4
pytz
APScheduler
